update web course list and semester automatically
