+++ static/css/my_css.css
@@ -15,7 +15,7 @@
 
 .chosen-drop {
   background-color: #3a3b45 !important;
-
+  box-shadow: 0px 5px 20px 1px black !important;
   border-radius: 25px !important;
   border: 0 !important;
   color: white !important;
@@ -36,18 +36,21 @@
 }
 
 .search-choice {
-  border-radius: 25px !important;
-  height: calc(1.5em + 0.75rem) !important;
+  border: 0 !important;
+  border-radius: 20px !important;
+  height: calc(1.5em + 0.5rem) !important;
   display: flex !important;
   align-items: center !important;
   justify-content: space-between !important;
   text-overflow: ellipsis;
   max-width: 200px;
+  margin: 5px 5px 5px 0 !important;
+  padding-left: 8px !important;
 }
 
 .search-choice span {
   text-overflow: ellipsis;
-  max-width: 8rem;
+  max-width: 8.5rem;
   white-space: nowrap;
   overflow: hidden;
   line-height: 1.5 !important;
